app rename complete
