+++ example/android/app/build.gradle
@@ -128,7 +128,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "me.kiano.rnbackgroundgeofencing"
+        applicationId "com.okhi.okhivaas"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
