+++ example/android/app/src/main/java/com/okhi/okhivaas/MainActivity.java
@@ -1,4 +1,4 @@
-package me.kiano.rnbackgroundgeofencing;
+package com.okhi.okhivaas;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "RNBackgroundGeofencing";
+    return "OkHiVaaS";
   }
 }
