+++ example/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">RNBackground Geofencing</string>
+    <string name="app_name">OkHi VaaS</string>
 </resources>
