+++ example/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "RNBackgroundGeofencing",
-  "displayName": "RNBackground Geofencing"
+  "name": "OkHiVaaS",
+  "displayName": "OkHi VaaS"
 }
