+++ example/ios/OkHiVaaS.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:RNBackgroundGeofencing.xcodeproj">
+      location = "group:OkHiVaaS.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
