+++ example/ios/OkHiVaaS/Images.xcassets/AppIcon.appiconset/Contents.json
