+++ example/ios/OkHiVaaS/Images.xcassets/Contents.json
