+++ example/ios/OkHiVaaSTests/Info.plist
