+++ example/ios/OkHiVaaSTests/OkHiVaaSTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface RNBackgroundGeofencingTests : XCTestCase
+@interface OkHiVaaSTests : XCTestCase
 
 @end
 
-@implementation RNBackgroundGeofencingTests
+@implementation OkHiVaaSTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
