+++ example/ios/Podfile
@@ -1,8 +1,8 @@
 platform :ios, '9.0'
 require_relative '../node_modules/@react-native-community/cli-platform-ios/native_modules'
 
-target 'RNBackgroundGeofencing' do
-  # Pods for RNBackgroundGeofencing
+target 'OkHiVaaS' do
+  # Pods for OkHiVaaS
   pod 'FBLazyVector', :path => "../node_modules/react-native/Libraries/FBLazyVector"
   pod 'FBReactNativeSpec', :path => "../node_modules/react-native/Libraries/FBReactNativeSpec"
   pod 'RCTRequired', :path => "../node_modules/react-native/Libraries/RCTRequired"
@@ -34,7 +34,7 @@ target 'RNBackgroundGeofencing' do
   pod 'glog', :podspec => '../node_modules/react-native/third-party-podspecs/glog.podspec'
   pod 'Folly', :podspec => '../node_modules/react-native/third-party-podspecs/Folly.podspec'
 
-  target 'RNBackgroundGeofencingTests' do
+  target 'OkHiVaaSTests' do
     inherit! :search_paths
     # Pods for testing
   end
@@ -42,10 +42,10 @@ target 'RNBackgroundGeofencing' do
   use_native_modules!
 end
 
-target 'RNBackgroundGeofencing-tvOS' do
-  # Pods for RNBackgroundGeofencing-tvOS
+target 'OkHiVaaS-tvOS' do
+  # Pods for OkHiVaaS-tvOS
 
-  target 'RNBackgroundGeofencing-tvOSTests' do
+  target 'OkHiVaaS-tvOSTests' do
     inherit! :search_paths
     # Pods for testing
   end
