+++ example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "RNBackgroundGeofencing",
+  "name": "OkHiVaaS",
   "version": "0.0.1",
   "private": true,
   "scripts": {
